Factor out common code for shapes.
