+++ com.ibm.wala.cast.python.ml/source/com/ibm/wala/cast/python/ml/client/TensorGenerator.java
@@ -3,13 +3,18 @@ package com.ibm.wala.cast.python.ml.client;
 import static com.ibm.wala.cast.python.ml.types.TensorFlowTypes.DType.FLOAT32;
 import static com.ibm.wala.cast.python.ml.types.TensorFlowTypes.D_TYPE;
 import static com.ibm.wala.cast.python.ml.types.TensorFlowTypes.TENSORFLOW;
+import static com.ibm.wala.cast.python.types.PythonTypes.list;
+import static com.ibm.wala.cast.python.util.Util.getAllocationSiteInNode;
 import static com.ibm.wala.core.util.strings.Atom.findOrCreateAsciiAtom;
 import static com.ibm.wala.ipa.callgraph.propagation.cfa.CallStringContextSelector.CALL_STRING;
+import static java.util.Arrays.asList;
 
+import com.ibm.wala.cast.ipa.callgraph.AstPointerKeyFactory;
 import com.ibm.wala.cast.python.ml.types.TensorFlowTypes;
 import com.ibm.wala.cast.python.ml.types.TensorFlowTypes.DType;
 import com.ibm.wala.cast.python.ml.types.TensorType;
 import com.ibm.wala.cast.python.ml.types.TensorType.Dimension;
+import com.ibm.wala.cast.python.ml.types.TensorType.NumericDim;
 import com.ibm.wala.cast.python.types.PythonTypes;
 import com.ibm.wala.classLoader.IClass;
 import com.ibm.wala.classLoader.IField;
@@ -17,6 +22,8 @@ import com.ibm.wala.classLoader.IMethod;
 import com.ibm.wala.classLoader.NewSiteReference;
 import com.ibm.wala.ipa.callgraph.CGNode;
 import com.ibm.wala.ipa.callgraph.ContextItem;
+import com.ibm.wala.ipa.callgraph.propagation.AllocationSiteInNode;
+import com.ibm.wala.ipa.callgraph.propagation.ConstantKey;
 import com.ibm.wala.ipa.callgraph.propagation.InstanceKey;
 import com.ibm.wala.ipa.callgraph.propagation.PointerAnalysis;
 import com.ibm.wala.ipa.callgraph.propagation.PointerKey;
@@ -67,23 +74,158 @@ public abstract class TensorGenerator {
     return ret;
   }
 
+  protected Set<List<Dimension<?>>> getShapes(
+      PropagationCallGraphBuilder builder, Iterable<InstanceKey> pointsToSet) {
+    Set<List<Dimension<?>>> ret = HashSetFactory.make();
+    PointerAnalysis<InstanceKey> pointerAnalysis = builder.getPointerAnalysis();
+
+    for (InstanceKey instanceKey : pointsToSet) {
+      AllocationSiteInNode asin = getAllocationSiteInNode(instanceKey);
+      TypeReference reference = asin.getConcreteType().getReference();
+
+      if (reference.equals(list)) { // TODO: This can also be a tuple of tensors.
+        // We have a list of integers that represent the shape.
+        OrdinalSet<InstanceKey> objectCatalogPointsToSet =
+            pointerAnalysis.getPointsToSet(
+                ((AstPointerKeyFactory) builder.getPointerKeyFactory())
+                    .getPointerKeyForObjectCatalog(asin));
+
+        // We expect the object catalog to contain a list of integers. Each element in the array
+        // corresponds to the set of possible dimensions for that index.
+        @SuppressWarnings("unchecked")
+        Set<Dimension<Integer>>[] possibleDimensions = new Set[objectCatalogPointsToSet.size()];
+
+        for (InstanceKey catalogIK : objectCatalogPointsToSet) {
+          ConstantKey<?> constantKey = (ConstantKey<?>) catalogIK;
+          Object constantKeyValue = constantKey.getValue();
+
+          Integer fieldIndex = (Integer) constantKeyValue;
+
+          FieldReference subscript =
+              FieldReference.findOrCreate(
+                  PythonTypes.Root, findOrCreateAsciiAtom(fieldIndex.toString()), PythonTypes.Root);
+
+          IField f = builder.getClassHierarchy().resolveField(subscript);
+          LOGGER.fine("Found field: " + f);
+
+          // We can now get the pointer key for the instance field.
+          PointerKey pointerKeyForInstanceField = builder.getPointerKeyForInstanceField(asin, f);
+          LOGGER.fine("Found pointer key for instance field: " + pointerKeyForInstanceField + ".");
+
+          // Get the points-to set for the instance field.
+          OrdinalSet<InstanceKey> instanceFieldPointsToSet =
+              pointerAnalysis.getPointsToSet(pointerKeyForInstanceField);
+          LOGGER.fine("Points-to set for instance field: " + instanceFieldPointsToSet + ".");
+
+          // If the instance field points to a constant, we can use it as the shape.
+          // TODO: Is it possible to also do it for (simple) expressions?
+          Set<Dimension<Integer>> tensorDimensions = HashSetFactory.make();
+
+          for (InstanceKey instanceFieldIK : instanceFieldPointsToSet) {
+            if (instanceFieldIK instanceof ConstantKey) {
+              // We have a constant key.
+              ConstantKey<?> instanceFieldConstant = (ConstantKey<?>) instanceFieldIK;
+              Object instanceFieldValue = instanceFieldConstant.getValue();
+
+              // We have a shape value.
+              Long shapeValue = (Long) instanceFieldValue;
+              LOGGER.fine(
+                  "Found shape value: " + shapeValue + " for " + source.getPointerKey() + ".");
+
+              Dimension<Integer> dimension = new NumericDim(shapeValue.intValue());
+
+              LOGGER.fine("Adding dimension: " + dimension + ".");
+              tensorDimensions.add(dimension);
+            } else
+              throw new IllegalStateException(
+                  "Expected a constant key for instance field: "
+                      + pointerKeyForInstanceField
+                      + ", but got: "
+                      + instanceFieldIK
+                      + ".");
+          }
+
+          LOGGER.info(
+              "Found possible shape dimensions: "
+                  + tensorDimensions
+                  + " for field: "
+                  + pointerKeyForInstanceField
+                  + " for source: "
+                  + source
+                  + ".");
+
+          // Add the shape dimensions.
+          assert possibleDimensions[fieldIndex] == null
+              : "Duplicate field index: "
+                  + fieldIndex
+                  + " in object catalog: "
+                  + objectCatalogPointsToSet
+                  + ".";
+
+          possibleDimensions[fieldIndex] = tensorDimensions;
+          LOGGER.fine(
+              "Added shape dimensions: "
+                  + tensorDimensions
+                  + " for field index: "
+                  + fieldIndex
+                  + ".");
+        }
+
+        for (int i = 0; i < possibleDimensions.length; i++)
+          for (Dimension<Integer> iDim : possibleDimensions[i]) {
+            @SuppressWarnings("unchecked")
+            Dimension<Integer>[] dimensions = new Dimension[possibleDimensions.length];
+
+            dimensions[i] = iDim;
+
+            for (int j = 0; j < possibleDimensions.length; j++)
+              if (i != j)
+                for (Dimension<Integer> jDim : possibleDimensions[j]) dimensions[j] = jDim;
+
+            ret.add(asList(dimensions));
+          }
+      } else
+        throw new IllegalStateException(
+            "Expected a " + PythonTypes.list + " for the shape, but got: " + reference + ".");
+    }
+
+    return ret;
+  }
+
+  protected abstract Set<List<Dimension<?>>> getDefaultShapes(PropagationCallGraphBuilder builder);
+
+  protected abstract int getValueNumberForShapeArgument();
+
   /**
    * Returns the possible shapes of the tensor returned by this generator.
    *
    * @param builder The {@link PropagationCallGraphBuilder} used to build the call graph.
    * @return a set of shapes, where each shape is represented as a list of dimensions
    */
-  protected abstract Set<List<Dimension<?>>> getShapes(PropagationCallGraphBuilder builder);
+  protected Set<List<Dimension<?>>> getShapes(PropagationCallGraphBuilder builder) {
+    PointerAnalysis<InstanceKey> pointerAnalysis = builder.getPointerAnalysis();
 
-  protected abstract EnumSet<DType> getDefaultDTypes(PropagationCallGraphBuilder builder);
+    // Get the shape from the explicit argument.
+    // FIXME: Handle keyword arguments.
+    int valueNumber = this.getValueNumberForShapeArgument();
+
+    PointerKey pointerKey = pointerAnalysis.getHeapModel().getPointerKeyForLocal(node, valueNumber);
+    OrdinalSet<InstanceKey> pointsToSet = pointerAnalysis.getPointsToSet(pointerKey);
+
+    // If the argument shape is not specified.
+    if (pointsToSet.isEmpty()) return getDefaultShapes(builder);
+    else
+      // The shape points-to set is non-empty, meaning that the shape was explicitly set.
+      return getShapes(builder, pointsToSet);
+  }
 
   protected EnumSet<DType> getDTypes(
-      PropagationCallGraphBuilder builder, Iterable<InstanceKey> dTypePointsToSet) {
+      PropagationCallGraphBuilder builder, Iterable<InstanceKey> pointsToSet) {
     EnumSet<DType> ret = EnumSet.noneOf(DType.class);
     PointerAnalysis<InstanceKey> pointerAnalysis = builder.getPointerAnalysis();
 
-    for (InstanceKey dTypeIK : dTypePointsToSet) {
-      IClass concreteType = dTypeIK.getConcreteType();
+    for (InstanceKey instanceKey : pointsToSet) {
+      IClass concreteType = instanceKey.getConcreteType();
       TypeReference typeReference = concreteType.getReference();
 
       if (typeReference.equals(TensorFlowTypes.D_TYPE)) {
@@ -139,7 +281,7 @@ public abstract class TensorGenerator {
                 .getPointerKeyForInstanceField(tensorFlowIK, float32Field);
 
         for (InstanceKey float32IK : pointerAnalysis.getPointsToSet(float32PK))
-          if (float32IK.equals(dTypeIK)) {
+          if (float32IK.equals(instanceKey)) {
             ret.add(FLOAT32);
             LOGGER.info(
                 "Found dtype: "
@@ -147,9 +289,9 @@ public abstract class TensorGenerator {
                     + " for source: "
                     + source
                     + " from dType: "
-                    + dTypeIK
+                    + instanceKey
                     + ".");
-          } else throw new IllegalStateException("Unknown dtype: " + dTypeIK + ".");
+          } else throw new IllegalStateException("Unknown dtype: " + instanceKey + ".");
       } else
         throw new IllegalStateException(
             "Expected a "
@@ -161,6 +303,8 @@ public abstract class TensorGenerator {
 
     return ret;
   }
+
+  protected abstract EnumSet<DType> getDefaultDTypes(PropagationCallGraphBuilder builder);
 
   protected EnumSet<DType> getDTypes(PropagationCallGraphBuilder builder) {
     PointerAnalysis<InstanceKey> pointerAnalysis = builder.getPointerAnalysis();
